+++ CHANGELOG.md
@@ -11,6 +11,26 @@ For more details or to discuss releases, please visit the
 
 ## [Unreleased]
 
+## Next
+
+## [[0.18.4] - 2025-08-04](https://github.com/simpleiot/simpleiot/releases/tag/v0.18.4)
+
+- don't return power information for Shelly devices that don't have power
+  measurement (like Plus 1)
+- support status for Shelly gen1 relays
+- update Go packages to
+  [latest versions](https://github.com/simpleiot/simpleiot/pull/752/commits/6f6a00ed4b27d0809ef28e96216631a2e8da9559)
+- fix bug with importing a list of nodes
+- add browser client allowing control and configuration of Yoe Kiosk Browser
+
+## [[0.18.3] - 2025-03-20](https://github.com/simpleiot/simpleiot/releases/tag/v0.18.3)
+
+- add favicon to frontend so icon displays in browser tabs (#756)
+
+## [[0.18.2] - 2025-03-19](https://github.com/simpleiot/simpleiot/releases/tag/v0.18.2)
+
+- fix bug where export/import displayed no nodes (#749). Fixed in (#753)
+
 ## [[0.18.1] - 2024-11-19](https://github.com/simpleiot/simpleiot/releases/tag/v0.18.1)
 
 - fix bug where Shelly IO Enable Control option was not working (#739)